requirements.txt
python-dotenv
crewai
textblob
newsapi
firecrawl
vaderSentiment